Esception Handling

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -1,14 +1,20 @@
 package com.devsuperior.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.dscommerce.dtos.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -18,7 +24,8 @@ public class ProductService {
 	
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Product product = productRepository.findById(id).get();
+		Product product = productRepository.findById(id).orElseThrow(
+				() -> new ResourceNotFoundException("Recurso não encontrado"));
 		return new ProductDTO(product);
 		
 	}
@@ -40,16 +47,27 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product p = productRepository.getReferenceById(id);
-		copyDtoToEntity(dto, p);		
-		p = productRepository.save(p);
-		
-		return new ProductDTO(p);
+		try {
+			Product p = productRepository.getReferenceById(id);
+			copyDtoToEntity(dto, p);		
+			p = productRepository.save(p);
+			return new ProductDTO(p);
+		} catch(EntityNotFoundException r) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		productRepository.deleteById(id);
+		if (!productRepository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+	       	productRepository.deleteById(id);    		
+		}catch (DataIntegrityViolationException e) {
+	       	throw new DatabaseException("Falha de integridade referencial");
+	   	}
+
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product p) {
